adjust dcquota code location
